Face-Recognition-using-Facenet/requirements.txt
facenet-pytorch
torch
torchvision
opencv-python-headless
numpy
scikit-learn
pillow
